s2180, rework: rename files to AdCutoutResolver and VOSpaceCutoutResolver
