+++ caom2-tap/src/main/resources/CaomResolver.properties
@@ -3,20 +3,17 @@
 #
 # scheme:<class name of SchemeHandler impl>
 
-# basic AD resolver
-ad=ca.nrc.cadc.caom2.artifact.resolvers.AdResolver
-# AD resolver that handles cutouts
+
+# AD resolver
 ad=ca.nrc.cadc.caom2.artifact.resolvers.AdResolverWithCutouts
 
 gemini=ca.nrc.cadc.caom2.artifact.resolvers.GeminiResolver
 
 # MAST resolver that maps to CADC archive directory structure
 mast=ca.nrc.cadc.caom2.artifact.resolvers.CadcMastResolver
+
 # basic MAST resolver
 mast=ca.nrc.cadc.caom2.artifact.resolvers.MastResolver
-
-# basic VOSpace resolver
-vos=ca.nrc.cadc.caom2.artifact.resolvers.VOSpaceResolver
 
 # VOSpace resolver that handles cutouts
 vos=ca.nrc.cadc.caom2.artifact.resolvers.VOSpaceResolverWithCutouts
